refactor: eliminate code duplication and improve code quality
- Extract shared harness runner (runner.rs) eliminating ~85% duplication
  between HostHarness and KataHarness
- Split daemon.rs (238→122+128) and client.rs (208→159+59) under 200-line limit
- Merge identical is_fatal_token_error/is_fatal_request_error into
  is_fatal_api_error
- Delete dead ContainerBoot error variant
- Consolidate parse tests into parse_tests.rs, remove duplicated tests
- Migration 006: widen tokens_used and duration_ms from INTEGER to BIGINT
- Update Rust types i32→i64 for tokens_used/duration_ms across main-app
- Regenerate sqlx-data.json

+++ main-app/src/routes/jobs.rs
@@ -17,8 +17,8 @@ struct PollResponse {
 pub struct SubmitResultBody {
     pr_url: String,
     exit_code: i32,
-    tokens_used: i32,
-    duration_ms: i32,
+    tokens_used: i64,
+    duration_ms: i64,
 }
 
 pub async fn poll(state: web::Data<AppState>, auth: WorkerAuth) -> Result<HttpResponse, AppError> {
