+++ main-app/src/services/work_runs/model.rs
@@ -28,8 +28,8 @@ pub struct WorkRun {
     pub agents_md: String,
     pub result_pr_url: Option<String>,
     pub result_exit_code: Option<i32>,
-    pub tokens_used: Option<i32>,
-    pub duration_ms: Option<i32>,
+    pub tokens_used: Option<i64>,
+    pub duration_ms: Option<i64>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
